(maint) Clean up old workspace at the beginning of every run
This commit removes the specific directory deletes and instead deletes
the whole workspace at the start of every run, clearing out any
artifacts left over from the last time the workspace was used (Jenkins
reuses named directories as workspaces without deleting them
inbetween). This is in lieu of using the Workspace Cleanup plugin, which
I couldn't get to work. There are a couple of benefits of cleaning up at
the beginning of a run instead of the end:
1) if you need to go inspect the workspace after a job has finished,
it's still available.
2) no complicated logic around making sure the workspace gets cleaned
even if the job fails.